Polished UI
Fixed Icon not Rendereing, changed sevend day forecast date format, and fixed bug that wouldn't close the progress dialog

+++ app/src/main/java/com/example/android/simpleweather/models/Current.kt
@@ -28,7 +28,7 @@ data class Current(
 
     val weather: List<Weather?>?,
 
-    val windDeg: Int?,
+    val windDeg: Long?,
 
     val windSpeed: Double?
 ): Serializable
